Move constants out to own file
