+++ packages/ui/src/components/Command/Command.constants.ts
@@ -1,0 +1,7 @@
+export const COMMAND_ROUTES = {
+  AI: 'Supabase AI',
+  DOCS_SEARCH: 'Docs Search',
+  THEME: 'Theme',
+  AI_ASK_ANYTHING: 'Ask anything',
+  AI_RLS_POLICY: 'Help me create a RLS policy',
+}
