+++ packages/ui/src/components/Command/CommandMenuShortcuts.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+
 import { CommandShortcut } from './Command.utils'
 import { useCommandMenu } from './CommandMenuProvider'
